MNT: separate aero surface classes into different files
