+++ rocketpy/rocket/aero_surface/__init__.py
@@ -1,0 +1,6 @@
+from rocketpy.rocket.aero_surface.air_brakes import AirBrakes
+from rocketpy.rocket.aero_surface.aero_surface import AeroSurface
+from rocketpy.rocket.aero_surface.tail import Tail
+from rocketpy.rocket.aero_surface.fins import Fins, TrapezoidalFins, EllipticalFins
+from rocketpy.rocket.aero_surface.nose_cone import NoseCone
+from rocketpy.rocket.aero_surface.rail_buttons import RailButtons
